requirements.txt
PyYAML>=6.0
watchdog>=3.0
pywin32>=306
pystray>=0.19
Pillow>=10.0
pyautogui>=0.9.54
pydirectinput>=1.0.4
